fix: Cleaned up the function names, and separated logic into separate files
In Go Lang, it seems that a package's functions should (or must) begin with a capital letter. I made this change, and also slightly changed the name to better describe the function. I also separated the two logic parts (Chromosome, and Population) into separate source files. This should improve readability and maintainability.