fix: calculate cost of each new Chromosome during Crossover phase
During a Population's Crossover phase, Chromosomes are paired up and Mate()-d. The Mate() method does just that: mates the two Chromosomes, but does not perform a cost function. It doesn't appear to have been an issue, but, for the sake of safety, a Crossover() method should return a working Population -- one in which its Cost has been calculated.